+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Fixed
+
+- Automatic Persisted Queries work with hybrid cache mode https://github.com/nuwave/lighthouse/pull/2727
+
 ## v6.63.1
 
 ### Fixed